[feat]: UINavigationController popToRootViewController handler 포함 버전 추가
