+++ Zatch/global/Resource/extenstion/UINavigationControllerViewController.swift
@@ -1,0 +1,19 @@
+//
+//  UINavigationControllerViewController.swift
+//  Zatch
+//
+//  Created by 박지윤 on 2022/09/08.
+//
+
+import UIKit
+
+extension UINavigationController {
+    
+    func popToRootViewController(animated: Bool = true, completion: @escaping () -> Void) {
+        CATransaction.begin()
+        CATransaction.setCompletionBlock(completion)
+        popToRootViewController(animated: animated)
+        CATransaction.commit()
+    }
+    
+}
